+++ packages/create-invoice-form/src/lib/create-invoice-form.svelte
@@ -1,7 +1,7 @@
 <svelte:options customElement="create-invoice-form" />
 
 <script lang="ts">
-  import { getAccount } from "@wagmi/core";
+  import { getAccount, watchAccount, WatchAccountReturnType } from "@wagmi/core";
   import { Config as WagmiConfig } from "wagmi";
   // Types
   import { type GetAccountReturnType } from "@wagmi/core";
@@ -23,15 +23,20 @@
   import Button from "@requestnetwork/shared-components/button.svelte";
   import Status from "@requestnetwork/shared-components/status.svelte";
   import Modal from "@requestnetwork/shared-components/modal.svelte";
-  import { EncryptionTypes, CipherProviderTypes } from '@requestnetwork/types';
+  import { EncryptionTypes, CipherProviderTypes } from "@requestnetwork/types";
+    import { onDestroy, onMount, tick } from "svelte";
+
+  interface CipherProvider extends CipherProviderTypes.ICipherProvider {
+    disconnectWallet: () => void;
+  }
 
   export let config: IConfig;
   export let wagmiConfig: WagmiConfig;
   export let requestNetwork: RequestNetwork | null | undefined;
   export let currencies: CurrencyTypes.CurrencyInput[] = [];
-  let cipherProvider: CipherProviderTypes.ICipherProvider | undefined = requestNetwork?.getCipherProvider();
+  let cipherProvider: CipherProvider | undefined;
 
-  let account: GetAccountReturnType;
+  let account: GetAccountReturnType | undefined = wagmiConfig && getAccount(wagmiConfig);
   let isTimeout = false;
   let activeConfig = config ? config : defaultConfig;
   let mainColor = activeConfig.colors.main;
@@ -115,11 +120,40 @@
     totalAmount: 0,
   };
 
-  $: {
-    if (wagmiConfig) {
-      account = getAccount(wagmiConfig);
+  $: cipherProvider = requestNetwork?.getCipherProvider() as CipherProvider;
+
+  const handleWalletConnection = async () => {
+    account = getAccount(wagmiConfig);
+  };
+
+  const handleWalletDisconnection = () => {
+    cipherProvider?.disconnectWallet();
+  };
+
+  const handleWalletChange = (data: any) => {
+    if (data?.address) {
+      handleWalletConnection();
+    } else {
+      handleWalletDisconnection();
     }
-  }
+  };
+
+  onMount(() => {
+    unwatchAccount = watchAccount(wagmiConfig, {
+      onChange(data) {
+        tick().then(() => {
+          console.log("Wallet changed");
+          handleWalletChange(data);
+        });
+      },
+    });
+  });
+
+  let unwatchAccount: WatchAccountReturnType | undefined;
+
+  onDestroy(() => {
+    if (typeof unwatchAccount === "function") unwatchAccount();
+  });
 
   $: {
     formData.creatorId = (account?.address ?? "") as string;
@@ -186,7 +220,7 @@
       try {
         addToStatus(APP_STATUS.PERSISTING_TO_IPFS);
         let request;
-        if(cipherProvider && formData.isEncrypted) {
+        if (cipherProvider && formData.isEncrypted) {
           const payeeEncryptionParams = {
             key: requestCreateParameters.requestInfo.payee?.value!,
             method: EncryptionTypes.METHOD.KMS,
@@ -203,7 +237,7 @@
               paymentNetwork: requestCreateParameters.paymentNetwork,
               contentData: requestCreateParameters.contentData,
             },
-            [payeeEncryptionParams, payerEncryptionParams],
+            [payeeEncryptionParams, payerEncryptionParams]
           );
         } else {
           request = await requestNetwork.createRequest({