+++ packages/invoice-dashboard/src/lib/view-requests.svelte
@@ -16,6 +16,7 @@
   import TxType from "@requestnetwork/shared-components/tx-type.svelte";
   import DashboardSkeleton from "@requestnetwork/shared-components/dashboard-skeleton.svelte";
   import { toast } from "svelte-sonner";
+  import Modal from "@requestnetwork/shared-components/modal.svelte";
   // Icons
   import ChevronDown from "@requestnetwork/shared-icons/chevron-down.svelte";
   import ChevronLeft from "@requestnetwork/shared-icons/chevron-left.svelte";
@@ -41,26 +42,42 @@
   import { CurrencyManager } from "@requestnetwork/currency";
   import { onDestroy, onMount, tick } from "svelte";
   import { formatUnits } from "viem";
-  import { debounce, formatAddress } from "../utils";
+  import { debounce, formatAddress, getEthersSigner } from "../utils";
   import { Drawer, InvoiceView } from "./dashboard";
   import { getPaymentNetworkExtension } from "@requestnetwork/payment-detection";
   import { CipherProviderTypes, CurrencyTypes } from "@requestnetwork/types";
-    import { checkStatus } from "@requestnetwork/shared-utils/checkStatus";
+  import { checkStatus } from "@requestnetwork/shared-utils/checkStatus";
+  import { ethers } from "ethers";
+
+  interface CipherProvider extends CipherProviderTypes.ICipherProvider {
+    getSessionSignatures: (
+      signer: ethers.Signer,
+      walletAddress: `0x${string}`,
+      domain: string,
+      statement: string
+    ) => Promise<any>;
+    disconnectWallet: () => void;
+  }
 
   export let config: IConfig;
   export let wagmiConfig: WagmiConfig;
   export let requestNetwork: RequestNetwork | null | undefined;
   export let currencies: CurrencyTypes.CurrencyInput[] = [];
 
-  let cipherProvider: CipherProviderTypes.ICipherProvider | undefined = requestNetwork?.getCipherProvider();
+  let cipherProvider: CipherProvider | undefined;
 
-  let sliderValueForDecryption = cipherProvider?.isDecryptionEnabled() ? "on" : "off";
+  let sliderValueForDecryption = JSON.parse(
+    localStorage?.getItem("isDecryptionEnabled") ?? "false"
+  )
+    ? "on"
+    : "off";
 
   let signer: `0x${string}` | undefined;
   let activeConfig = config ? config : defaultConfig;
   let mainColor = activeConfig.colors.main;
   let secondaryColor = activeConfig.colors.secondary;
-  let account: GetAccountReturnType = wagmiConfig && getAccount(wagmiConfig);
+  let account: GetAccountReturnType | undefined =
+    wagmiConfig && getAccount(wagmiConfig);
 
   let loading = false;
   let searchQuery = "";
@@ -75,8 +92,7 @@
       })
     | undefined;
   let currencyManager: CurrencyManager;
-  let previousWalletAddress: string | undefined;
-  let previousNetwork: string | undefined;
+  let loadSessionSignatures = false;
 
   let columns = {
     issuedAt: false,
@@ -90,43 +106,44 @@
   let sortOrder = "desc";
   let sortColumn = "timestamp";
 
-  $: {
-    if (wagmiConfig) {
-      account = getAccount(wagmiConfig);
-    }
-  }
+  const handleWalletConnection = async () => {
+    account = getAccount(wagmiConfig);
+    await loadRequests(sliderValueForDecryption, account, requestNetwork);
+  };
 
-  $: {
-    if (account?.address) {
-      tick().then(() => {
-        getRequests();
-      });
-    }
-  }
+  const handleWalletDisconnection = () => {
+    account = undefined;
+    requests = [];
+    activeRequest = undefined;
+    cipherProvider?.disconnectWallet();
+    cipherProvider = undefined;
+  };
 
-  let unwatchAccount: WatchAccountReturnType | undefined;
+  const handleWalletChange = (data: any) => {
+    if (data?.address) {
+      handleWalletConnection();
+    } else {
+      handleWalletDisconnection();
+    }
+  };
 
   onMount(() => {
     unwatchAccount = watchAccount(wagmiConfig, {
       onChange(data) {
-        if (data?.address !== account?.address) {
-          account = data;
-          if (account?.address) {
-            getRequests();
-          } else {
-            requests = [];
-            activeRequest = undefined;
-            previousWalletAddress = undefined;
-            previousNetwork = undefined;
-          }
-        }
+        tick().then(() => {
+          handleWalletChange(data);
+        });
       },
     });
   });
 
+  let unwatchAccount: WatchAccountReturnType | undefined;
+
   onDestroy(() => {
     if (typeof unwatchAccount === "function") unwatchAccount();
   });
+
+  $: cipherProvider = requestNetwork?.getCipherProvider() as CipherProvider;
 
   $: {
     signer = account?.address;
@@ -138,9 +155,11 @@
     currencyManager = initializeCurrencyManager(currencies);
   });
 
-  const getRequests = async () => {
+  const getRequests = async (
+    account: GetAccountReturnType,
+    requestNetwork: RequestNetwork | undefined | null
+  ) => {
     if (!account?.address || !requestNetwork) return;
-    loading = true;
 
     try {
       const requestsData = await requestNetwork?.fromIdentity({
@@ -152,8 +171,6 @@
         .sort((a, b) => b.timestamp - a.timestamp);
     } catch (error) {
       console.error("Failed to fetch requests:", error);
-    } finally {
-      loading = false;
     }
   };
 
@@ -178,25 +195,6 @@
   const itemsPerPage = 10;
   let currentPage = 1;
   let totalPages = 1;
-
-  $: {
-    const currentWalletAddress = account?.address;
-    const currentNetwork = account?.chainId?.toString();
-
-    if (
-      currentWalletAddress &&
-      currentWalletAddress !== previousWalletAddress
-    ) {
-      getRequests();
-      previousWalletAddress = currentWalletAddress;
-
-      activeRequest = undefined;
-    }
-
-    if (currentNetwork && currentNetwork !== previousNetwork) {
-      previousNetwork = currentNetwork;
-    }
-  }
 
   $: {
     if (sortColumn && sortOrder) {
@@ -329,7 +327,7 @@
           BigInt(request.expectedAmount),
           currencyInfo?.decimals ?? 18
         ),
-        currencySymbol: currencyInfo!.symbol,
+        currencySymbol: currencyInfo?.symbol ?? "-",
         paymentCurrencies,
       };
     }
@@ -390,23 +388,66 @@
     activeRequest = undefined;
   };
 
-  
-  $: sliderValueForDecryption, getRequests(); 
+  const loadRequests = async (
+    sliderValue: string,
+    currentAccount: GetAccountReturnType | undefined,
+    currentRequestNetwork: RequestNetwork | undefined | null
+  ) => {
+    if (!currentAccount?.address || !currentRequestNetwork || !cipherProvider)
+      return;
 
-  $: {
-    if(sliderValueForDecryption === 'on') {
-      cipherProvider?.enableDecryption(true);
+    loading = true;
+    if (sliderValue === "on") {
+      try {
+        const signer = await getEthersSigner(wagmiConfig);
+        if (signer && currentAccount?.address) {
+          loadSessionSignatures =
+            localStorage?.getItem("lit-wallet-sig") === null;
+          await cipherProvider?.getSessionSignatures(
+            signer,
+            currentAccount.address,
+            window.location.host,
+            "Sign in to Lit Protocol through Request Network"
+          );
+          cipherProvider?.enableDecryption(true);
+          localStorage?.setItem("isDecryptionEnabled", JSON.stringify(true));
+        }
+      } catch (error) {
+        console.error("Failed to enable decryption:", error);
+        toast.error("Failed to enable decryption.");
+        loading = false;
+        return;
+      } finally {
+        loadSessionSignatures = false;
+      }
     } else {
       cipherProvider?.enableDecryption(false);
+      localStorage?.setItem("isDecryptionEnabled", JSON.stringify(false));
     }
-  }
+    await getRequests(currentAccount, currentRequestNetwork);
+    loading = false;
+  };
 
+  $: loadRequests(sliderValueForDecryption, account, requestNetwork);
 </script>
 
 <div
   class="main-table"
   style="--mainColor: {mainColor}; --secondaryColor: {secondaryColor}; "
 >
+  {#if loadSessionSignatures}
+    <Modal {config} isOpen={true} title="Lit Protocol Signature Required">
+      <div class="modal-content">
+        <p>
+          This signature is required only once per session and will allow you
+          to:
+        </p>
+        <ul>
+          <li>Access encrypted invoice details</li>
+        </ul>
+      </div>
+    </Modal>
+  {/if}
   <div class="tabs">
     <ul>
       <li
@@ -443,11 +484,16 @@
         </Input>
         {#if cipherProvider}
           <div class="width: fit-content;">
-            <Switch bind:value={sliderValueForDecryption} label="Show encrypted requests" fontSize={14} design="slider" />
+            <Switch
+              bind:value={sliderValueForDecryption}
+              label="Show encrypted requests"
+              fontSize={14}
+              design="slider"
+            />
           </div>
         {/if}
       </div>
-      
+
       <Dropdown
         config={activeConfig}
         type="checkbox"
@@ -671,12 +717,16 @@
                 <td>
                   <TxType
                     type={signer === request.payer?.value ? "OUT" : "IN"}
+                    showBoth={request.payer?.value === request.payee?.value}
                   />
                 </td>
                 <td><StatusLabel status={checkStatus(request)} /></td>
                 <td>
                   {#if request.paymentCurrencies.length > 0}
-                    <Network network={request.paymentCurrencies[0]?.network} />
+                    <Network
+                      network={request.paymentCurrencies[0]?.network}
+                      showLabel={true}
+                    />
                   {:else}
                     <span class="text-gray-400">-</span>
                   {/if}
@@ -711,10 +761,8 @@
                 >
               </tr>
             {/each}
-          {:else}
-            {#if loading}
-              <DashboardSkeleton />
-            {/if}
+          {:else if loading}
+            <DashboardSkeleton />
           {/if}
         </tbody>
       </table>
@@ -820,7 +868,7 @@
   }
 
   .tabs ul li {
-    width: 100px;
+    width: 110px;
     display: inline-flex;
     align-items: center;
     justify-content: center;
@@ -1024,5 +1072,20 @@
     color: black;
     font-size: 20px;
     margin-bottom: 0;
+  }
+
+  .modal-content {
+    padding: 1rem;
+  }
+
+  .modal-content ul {
+    list-style-type: disc;
+    margin-left: 1.5rem;
+    margin-top: 0.5rem;
+  }
+
+  .modal-content li {
+    margin-bottom: 0.5rem;
+    color: #4b5563;
   }
 </style>