+++ shared/utils/getNetworkIcon.ts
@@ -14,6 +14,7 @@ import OptimismIcon from "../icons/network/optimism.svelte";
 import ArbitrumIcon from "../icons/network/arbitrum.svelte";
 import MoonbeamIcon from "../icons/network/moonbeam.svelte";
 import FantomIcon from "../icons/network/fantom.svelte";
+import Unknown from "../icons/unkown.svelte";
 
 export const getNetworkIcon = (network: string) => {
   const icons = {
@@ -35,5 +36,5 @@ export const getNetworkIcon = (network: string) => {
     moonbeam: MoonbeamIcon,
   };
 
-  return icons[network] || EthereumIcon;
+  return icons[network] || Unknown;
 };