+++ packages/create-invoice-form/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @requestnetwork/create-invoice-form
 
+## 0.11.12
+
+### Patch Changes
+
+- Fix Lit Encryption and Wallet Connections
+
+## 0.11.11
+
+### Patch Changes
+
+- Fix Lit Protocol Encryption Issues
+
+## 0.11.10
+
+### Patch Changes
+
+- Fix enbable decryption
+
+## 0.11.9
+
+### Patch Changes
+
+- Upgrade SDK to 0.53.0
+
 ## 0.11.8
 
 ### Patch Changes