+++ packages/payment-widget/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @requestnetwork/payment-widget
 
+## 0.3.7
+
+### Patch Changes
+
+- Fix Lit Protocol Encryption Issues
+
+## 0.3.6
+
+### Patch Changes
+
+- Upgrade SDK to 0.53.0
+
 ## 0.3.5
 
 ### Patch Changes